Expose InvalidConfig & Unsupported errors to client, fix Setup signature
In the long run, we want both of these sent back to core, so it can inform the UI about whether there's an
error in the job spec or if the job spec says to use ChainReader but they're running an old relay that doesn't
support it.

While we're still migrating MedianProvider to using ChainReader, we can ignore the Unsupported error (either
on the relay side or core side?). But we should still return an error if they try to configure ChainReader
improperly.

If the ChainReader section of RelayConfig is completely missing from the job spec, we can initially return
Unsupported (treating this as the user intentionally wanting to disable ChainReader). Then after the migration,
return InvalidConfig--since this will be a required part of the RelayConfig.