Allow Value Checks on Static Provider
The static provider check incoming values against constants in the test package. For fuzz testing, this results in
errors. In the case of fuzz testing, static checks should not be necessary and instead the provider can simply return
the configured static values. The solution was to provide a configuration down the stack for static checks.