Bcf 2572 relayer fuzz test (#322)
* Create Fuzz Testing for Relayer

This commit adds basic fuzz testing functions that apply fuzzing to `PluginRelayer.NewRelayer`
`Relayer.NewConfigProvider`. Since fuzzing does not currently support parallel runs, the fuzz
functions are designed in a granular way where a test constructor function that returns the
testable interface is provided to the fuzz testing functions.

Nested fuzz tests have not yet been tried, but this generally appears to be a bad idea
considering how fuzz testing works. ex:

```
func RunFuzzPluginRelayer(f *testing.F) {
    f.Fuzz(func(t *testing.T)) {
        f.Fuzz(func(t *testing.T)) {

        }
    }
}
```

- Allow Value Checks on Static Provider

The static provider check incoming values against constants in the test package. For fuzz testing, this results in
errors. In the case of fuzz testing, static checks should not be necessary and instead the provider can simply return
the configured static values. The solution was to provide a configuration down the stack for static checks.

- fix linting issues