Replace Pending boolean with Confidence
This was being used to determine whether to use latest block or
finalized block for event reads (referred to as Finalized in the
CLIP, but renamed to Pending during implementation due to some
miscommunications around its purpose.

Having it be a Confidence level from 0 to 10 allows for more
granularity but still avoids products needing to pass a chain-specific # of
confirmations.  Instead, the Confidence level should be translated into # of
confirmations based on the chain, in the RelayConfig.

Choosing int32 instead of uint32 leaves the door open for special
values like Confidence = -1 which could be used to represent the Pending
block in the tx pool. (Presently this is not supported, and there
are no known use cases, but better to be more flexible in case the
need arises.)