Change Confirmations to Confidence
`Confidence` is a different approach to mapping a chain agnostic concept of confidence that
an event or transaction can or should be acted on. `Confidence` is intended to map to defined
levels of confirmations on EVM chains and other states on other chains.

Only two levels of `Confidence` exist by default to define the boundaries: `Highest` and
`Lowest`. This creates an inclusive range for all chain configurable `Confidence` levels and
can be treated as % confidence that an event or transaction can or should be acted on.