Send Metering Report to Beholder (#1062)
This commit adds a new message to the capabilities proto that defines a custom message for a metering report.

Additionally, a new test util was added to support setting the global beholder client with a data collector. This data
collector provides assertions on messages sent to beholder for use in unit tests.