Proposal for Custom Beholder Messages
The default custom message utility uses a `BaseMessage` proto with the message data expected to
be a string and labels added to the proto message.

Instead, a proto message itself can be sent where the message and schema details are packaged
together by an interface. This promotes the design that any particular entity can be emitted
given that it satisfies the interface.