Add Metering Detail to Capabilities Response
Every node must independently report metering information while the engine expects an aggregated list of all nodes.
Instead of having two different response types: one for a node and one as an aggregate, a `CapabilityResponse` can
contain metadata with an array of values. A node is expected to send a single entity in the array while an aggregated
response would have multiple.

This change gives the ability to surface metering data from 1 or many nodes without moving to more complex types.