Capability Interface Test Framework
This commit introduces a testing framework for asserting essential tests that all capabilities should conform to. As a
starting point, both trigger and executable capabilities must conform to the `BaseCapability` interface and produce
properly structured information. Additionally, executable capabilities should return metering information in
a `CapabilityResponse`.

Tests for the `BaseCapability` cannot be disabled but those for `ExecutableCapability` can be per capability
implementation as each implemtation will be different.