Split Spending Limits to Capability Types
Some capabilities can have multiple spending types while user defined spending limits or balance spending limts are in
universal credits. If a capability limit is set by the user as a single value of universal credits, we need to split
that limit and convert the split values into native units for a capability. The problem is the knowing the ratio ahead
of calling the capability for how resources should be split.

This commit introduces a new value in the `Info` call that returns a resource type and the maximum value of that unit
for that capability. From the max values, the engine can infer the ratio between resource types and split the user
defined limit on that ratio.